+++ src/main/java/com/gps/GuardPrivacySystem/repository/UtenteRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.gps.GuardPrivacySystem.model.Utente;
 
 @Repository
-public interface UtenteRepository extends JpaRepository<Utente, Long>{
+public interface UtenteRepository extends JpaRepository<Utente, Integer>{
 	
 }
